Release Fibers that are not added back to the pool.
This commit fixes issue https://github.com/meteor/meteor/issues/10359.
Fibers that are created and not returned to the pool are not
released. They will accumulate over time and the node process will get
stuck and must be restarted.

This change simple resets handle to a fiber when we are done with it...
which releases it properly.