Revert "Skip regenerating main bundle when only a dynamic file changed."
This reverts commit 95a1f54ccef068327c37dc5aba97b8b25189dcb7.

Although at first this commit seemed like a nice optimization, after
digging into #10496 I was reminded that the main bundle includes the paths
and hashes of available dynamic modules, so rebuilding the main bundle is
important even when only dynamic modules have changed.

@zodern I should have caught this in my review, sorry! Remembering how all
the pieces fit together is a challenge, to say the least.

Many thanks to @arggh for testing 1.8.1-beta.20 and putting together the
https://github.com/arggh/meteor-bundler-issue reproduction.