Only include direct npm dependencies of a Meteor package
In case a file within a package requires/imports an npm dependency
We should only resolve this if it is mentioned in the npm dependencies
of that package (thus in `Npm.depends` in the package.js file).
Otherwise, the package author likely wants to get it from the app's
node_modules.
The npm package could be in a node modules path of the meteor package
through some other npm package.
So we still want this other npm package to resolve to that package
Since its version may differ from the app's npm package