Add Storybook to the roadmap
Related to switching to NPM, it would be a good occasion to open up to a broader tooling ecosystem.

A big missing piece from the developer standpoint is Storybook. It's not just some nice tool, it literally changed the way we develop frontend in my company and all the companies where I've managed to introduce this magical tool.

Sadly it doesn't work correctly with Meteor, because it is based on a Webpack build. This issue haven't really been solved. We've added Storybook to Vulcan, but that's more a hack than an actual solution (basically we scrap Meteor imports + rely on Vulcan's naming convention to find package root...) so I am not personaly knowledgeable enough to help :/

Another great tool is Jest, it's way more efficient than Mocha thanks to its powerful CLI, but needs some webpack build.