Include Meteor.isSever constant in roadmap
I (and presumably others) often have code shared between client and server, with some blocked off by `if(Meteor.isServer)` and (more rarely) `if(!Meteor.isServer)`. It would be great if this constant is also handled in the tree shaking, and I imagine it's about as easy as `Meteor.isProduction` (but slightly different so worth mentioning).