Revert changes to webapp package since version 1.8.0.
We accidentally published changes to webapp that should have been
restricted to Meteor 1.10 as part of the 1.8.1 version. This commit
reverts commits to packages/webapp since Meteor 1.9, so that we can
republish the 1.8.0 content as version 1.8.2. We will then bump the webapp
version to 1.9.0 on the release-1.10 branch and publish the new content
only on that branch.

Revert "Allow to exclude web architectures in development mode (#10824)"
This reverts commit a20596718648727e384a6969ac2160f505aac763.

Revert "Updates cordova-plugin-meteor-webapp to 1.7.1"
This reverts commit a1e4d27822852e357bdebaffcaa15e6afe1ab94a.

Revert "Update cordova-plugin-wkwebview-engine to 1.2.1."
This reverts commit 3f9a69d7c454d832b0cbb7567ec47fe138304417.

Revert "Update cordova-plugin-whitelist to 1.3.4."
This reverts commit 979273333b438bcd8e6ff313f1b4c5cb304ea313.

Revert "Update cordova-plugin-meteor-webapp to 1.7.1-beta.1."
This reverts commit 565c4254f190c96b3cf51d762e721731aae061a0.

Revert "Update accounts-password to version 1.5.2."
This reverts commit b827d1da2f4434726a1fc7e8294b638ba054ee1a.