Gracefully handle EPIPE errors for nonessential IPC messages.
These stray EPIPE messages have been a problem for a while now, especially
on Windows.

In most cases, they appear to stem from sending a message to a process
that is about to exit, such that the message races against the exit, and
sometimes loses. When that message is a just an obligatory response to a
final message from the exiting process, the exiting process probably does
not care about (and will not receive) the response, so we can safely
swallow the EPIPE error.

Now that all otherProcess.send calls have a callback function, I'm hopeful
we will never see EPIPE errors again.