Add 'variables' as an extension for the less compiler to act on
In some projects, it's common to define `site.variables` or `theme.variables` less files. However, these do not appear to get picked up correctly by the less compiler (even though they are imported from `.less` files), so I added `variables` as a file extension for the less compiler to act on. So far, it seems like it works 😅
I suppose it would be even more ideal to have some config in the `package.json` for this, because some projects, like Semantic UI, use less files called `theme.config` and so on as well.

Feel free to close this if it's not relevant. I just wanted to make sure there was some record of this issue!