The worker process name that is used to name each worker process's UNIX socket file can now be overridden with the environment variable NAME.
This is useful because the worker id, which is used by default to name the UNIX socket file, is incremented each time the worker process ends/dies. This will cause difficulties if you are referencing the UNIX socket file from a web server or other external process where you expect the socket file to always have the same name.

Co-authored-by: Rokas Stankevicius <rokanost@users.noreply.github.com>