Add platforms options to meteor build:
This commit adds the "--platforms" flag to meteor build command. This allows the user to specify which platforms he want to build and build it only. For example, if we pass only android to it, we will build only android and web.cordova.

Related to issue #63 (https://github.com/meteor/meteor-feature-requests/issues/63)