fix: allow dynamic import to work when meteor is running with a path
When setting "useLocationOrigin: true" and also using a root url like
ROOT_URL=http://host:port/mypath , dynamic-import will ignore
"mypath" when fetching for files to be imported, resulting in a 404 for these
files.

To bypass this problem, i changed the code to prepend the "mypath" to the
module's URL, before trying to fetch it.

Related to #11105.