Add .gitattributes file
Added a .gitattributes file in the root of the project indicating text=auto so that Git clients will automatically convert line endings to the correct platform line endings when cloning, and normalize them back to LF when committing. Further, adding separate configurations for identified file formats to leave untouched as binary.