Update current version after every migration
Previously, if there were multiple pending migrations and a later
migration failed, the current version would still be set to before the
first migration of the series. That's inaccurate, since the earlier
migrations in the series did run successfully, and running those
migrations multiple times may not be idempotent (and is almost certainly
unexpected).

Instead, store the current version after each migration, but keep
control locked until the series finishes.