Permit users to set uname info
This allows users on multi-arch systems to explicitly set their architecture in their environment.

Overriding `ARCH` is necessary for systems which run without a `uname` binary which has been extended to "lie" about the system's architecture in different contexts. For example on a multi-arch build system `mongo` may be running in an execution mode which differs from the host execution mode; currently you either have to modify `mongo` or patch `uname` to "lie" in order to get the desired `ARCH`.

This change doesn't require a patched `mongo` or `uname` binary.
  Ex:
```
PATH="/usr/local/opt/gnu-coreutils/bin${PATH+:$PATH}" ARCH=x86_64;
export PATH ARCH;
/usr/bin/arch -x86_64 ~/.meteor/packages/meteor-tool/2.5.6/mt-os.osx.x86_64/meteor ARGS...;
```
For context `gnu-coreutils`'s `uname` is agnostic to the effects of `arch -x86_64` and would output `arm64` in this case when we actually want `x86_64`.