Fix errors in Mongo types
Both the synchronous and asynchronous iterators for Mongo.Cursor are
defined to have a "next" argument type of never. However, TypeScript
will reject using such an iterator in a for-of loop with:

> Cannot iterate value because the 'next' method of its iterator expects
> type 'never', but for-of will always send 'undefined'.ts(2763)

This changes back to the default (undefined), which is acceptable, since
the iterators ignore any arguments passed into next() anyway.

It also correctly declares the return type of dropIndexAsync, which
should return a Promise, not void.