OnlyOneOf support for match.js
Proposing to add `OnlyOneOf` as the equivalent of JSON Schema's [oneOf](https://json-schema.org/understanding-json-schema/reference/combining.html?highlight=oneof#oneof). It checks that there is one and only one match among the choices passed in. Could also be called `ExactlyOneOf` but I went with the shorter version.

I assumed that changing `Match.OneOf` to `Match.AnyOf`(to mimic JSON Schema's nomenclature) and using `Match.OneOf` for this new proposal would probably not be great for people who rely on this package.