Thread upstream mongodb types through npm-mongo
As it exists currently, the types for the mongo package don't work out
of the box because they can't find the types for the mongodb NPM
package. TypeScript app developers could work around this by adding
mongodb directly as a dependency, but they shouldn't have to.

However, `zodern:types` can handle pulling in types for NPM
dependencies, but only if the package _with_ those NPM dependencies is
itself exporting type declarations. Since npm-mongo is actually
depending on the mongodb package, add a type declaration reflecting both
the types and objects that it exports. Then switch the mongo package to
import from npm-mongo, giving it access to the types for the mongodb
package.