Fix: correct spelling of EACCES in rename
On windows and WSL, renaming a file can fail because of EPERM or EACCES, so there's a special version of the rename function that retries those errors up to a time limit, then copies the file and deletes the old one instead. However, the decision to retry is based on the correct name of the code (EACCES), and the decision to copy and delete is based on a misspelling (EACCES_S_). This fixes the typo.