Re-enable TypeScript self-test
The latest release of zodern:types is able to pull type declarations
from a Meteor checkout, if it's being used, instead of only from the
package repository (which may not contain the latest changes). This
means that we should be able to re-enable the TypeScript test, which
verifies that the TypeScript skeleton does, in fact, produce a project
which passes type checking.

(Re-enabling the test also picked up a few latent bugs in the type
declarations, so fix those as well)