Revert localhost change
This looks like it was a change initially done for some unrelated test and the docs went back and forth with specifying wether localhost should be by default replaced with 127.0.0.1
 https://github.com/meteor/meteor/commit/d056a26f0277b8e559cf32fc26be37b3638914cf

It seems like the logic that eventually got placed here is completely wrong, because it could replace a ROOT_URL with 127.0.0.1 which seems very wrong. 

I don't see why this should be a random breaking change here. so this change reverts it back to what it was, and only replace the url if it is localhost and the code specifies that it should