Set minPoolSize for oplog
Oplog has its specific startup, but it was missing a condition for minPoolSize which made it impossible to set minPoolSize for normal connection as setting it in the mongo settings would set it as well for oplog connection which would result in error as maxPoolSize for oplog is 1. So setting minPoolSize to 30 would result in failed startup on Galaxy when the oplog was started as minPoolSize for oplog could not be more than maxPoolSize (which is 1).