Update testing.md to show how to clean up Blaze View instances.
This adds an additional `Blaze.remove(view)` call to the sample test function to ensure the Blaze view is cleaned up properly. Additionally, this is great, even if not needed, because the Meteor AI in the new doc site answers questions about "how to render Blaze views with JavaScript" based on this `Blaze.renderWithData` example, so it is important for the AI to tell a person asking how to propertly clean up the `Blaze.View`.