Add type declarations for Assets
Assets isn't actually shipped in a Meteor package, but rather is
injected by the bundler as a global namespace, so there's not a great
place to add the type declarations. Given that the meteor package is
required for even a minimal Meteor app, adding the types there seems
like the best option.